Add files via upload
Two cars are located by using EVK-M8U and get the ground truth of GPS position of each car in the earth frame. One car is armed with a Inter D435 depth camera and take both RGB and Depth images. Yolo obejct detection NN is used to find out the desired car established with the GPS module. The Yolo ouput is the results6_nissan.txt and results.txt consists of boounding box pixel coordinate of the desired obejects. Pinhole camera model and depth image are used to estimate the 3D location of the desired car. Reults show the GPS location of two cars and the estimated position of the car through depth image.